Refactoring authorization/authentication moments
